pyproject.toml
[tool.poetry]
name = "django-chat"
version = "0.1.0"
description = "django_chat"
authors = ["Olga Tsuban <tsubanolga@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"


[tool.poetry.group.images.dependencies]
pillow = "^10.2.0"


[tool.poetry.group.test.dependencies]
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
